ai-travel-companion/requirements.txt
requests
google-api-python-client
python-dotenv
flask
gunicorn
